Enum megosabi::svc::Function

#[repr(u32)]
pub enum Function {

    Exit,
    PrintString,
    Monotonic,
    Time,
    Usleep,
    GetSystemInfo,
    NewWindow,
    CloseWindow,
    BeginDraw,
    EndDraw,
    DrawString,
    FillRect,
    DrawRect,
    DrawLine,
    Blt8,
    Blt1,
    Blt32,
    BlendRect,
    WaitChar,
    ReadChar,
    Rand,
    Srand,
    Alloc,
    Free,
    Test,

}

Variants

Exit

Terminate the process normally

PrintString

Display a string

Monotonic

Get the value of the monotonic timer in microseconds

Time

Perform the time service

Usleep

Blocks a thread for the specified microseconds

GetSystemInfo

Get the system information

NewWindow

Create a new window

CloseWindow

Close a window

BeginDraw

Create a drawing context

EndDraw

Discard the drawing context and reflect it to the screen

DrawString

Draw a string in a window

FillRect

Fill a rectangle in a window

DrawRect

Draw a rectangle in a window

DrawLine

Draw a line in a window

Blt8

Draw a bitmap in a window

Blt1

Draw a bitmap in a window

Blt32

Draw a bitmap in a window

BlendRect

Blend (test)

WaitChar

Wait for char event

ReadChar

Read a char event

Rand

Return a random number

Srand

Set the seed of the random number

Alloc

RESERVED

Free

RESERVED

Test

test_u64